Added $IF 32 and compression
Added an $IF 32BIT code block to automatically choose which open file dialog to use when someone compiles the code, making it work on both 32 and 64 bit for the open file dialogs. A problem that was present in many applications that use the COMDLG32 library as it is quite difficult to make work in 64 bit applications. Instead, I have written a PowerShell script that uses a very similar function to call an open file dialog and send it back to the application in a text file that is immediately deleted after being read. The Replace.BM file is only needed in a 64 bit compile as it is necessary for the functioning of the PowerShell script at runtime. Also, Dav, whose code I had been using in previous versions, has released a new algorithm for turning a binary file into a BAS file using the new _DEFLATE$ and _INFLATE$ functions which results in much smaller output files that can be $INCLUDEd in your programs. This version of my program uses that new algorithm and works faster because the input file is compressed. Many thanks to Dav and Fellippe on their encouragement!