Added MEM image function from Dav
Dav showed me a new function he designed to convert a picture into memory that can be called up to the QB64 executable's screen. I've adapted that code into my program to allow you to convert an image to MEM data so that the image never has to be written to the disk, only kept in a memory block.